2d particles

+++ src/vect2d.cpp
@@ -7,7 +7,7 @@ vect2d vect2d::operator+(vect2d const& obj){
     return value;
 }
 
-vect2d vect2d::operator*(float const &obj) {
+vect2d vect2d::operator*(double const &obj) {
     vect2d value;
     value.x = obj * x;
     value.y = obj * y;
@@ -38,29 +38,29 @@ vect2d vect2d::operator-(vect2d const &obj) {
     return value;
 }
 
-float vect2d::getAngle(){
+double vect2d::getAngle(){
     return atan(y/x);
 }
 
-float vect2d::getMag(){
+double vect2d::getMag(){
     return sqrt((x*x) + (y*y));
 }
 
-vect2d vect2d::rotate(float angle) {
-    float px = x * cos(angle) - y * sin(angle);
-    float py = x * sin(angle) + y * cos(angle);
+vect2d vect2d::rotate(double angle) {
+    double px = x * cos(angle) - y * sin(angle);
+    double py = x * sin(angle) + y * cos(angle);
     x = px;
     y = py;
     return *this;
 }
 
-vect2d::vect2d(float x, float y) {
+vect2d::vect2d(double x, double y) {
     this->x = x;
     this->y = y;
 }
 
 void vect2d::normalize(){
-    float mag = this->getMag();
+    double mag = this->getMag();
     x /= mag;
     y /= mag;
 }
